scripts/requirements.txt
certifi==2022.9.24
cffi==1.15.1
cryptography==38.0.1
cycler==0.11.0
fonttools==4.25.0
idna==3.4
kiwisolver==1.3.2
matplotlib==3.5.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
msgpack-python==0.4.8
munkres==1.1.4
numpy==1.21.5
packaging==21.3
Pillow==9.0.1
pip==21.2.4
pyasn1==0.4.8
pycparser==2.21
pymerkle==3.0.0
pyparsing==3.0.4
pysyncobj==0.3.11
python-dateutil==2.8.2
PyYAML==6.0
setuptools==65.6.0
six==1.16.0
tornado==6.1
tqdm==4.64.1
wheel==0.37.1
